Refine DCP doctor connection checker
