+++ src/Aspire.Cli/Program.cs
@@ -556,7 +556,7 @@ public class Program
         builder.Services.AddSingleton<IEnvironmentCheck, DeprecatedWorkloadCheck>();
         builder.Services.AddSingleton<IEnvironmentCheck, DevCertsCheck>();
         builder.Services.AddSingleton<IEnvironmentCheck, ContainerRuntimeCheck>();
-        builder.Services.AddSingleton<IDcpConnectionTester, DcpConnectionTester>();
+        builder.Services.AddSingleton<IDcpConnectionChecker, DcpConnectionChecker>();
         builder.Services.AddSingleton<IEnvironmentCheck, DcpConnectionHealthCheck>();
         builder.Services.AddSingleton<IEnvironmentCheck, DeprecatedAgentConfigCheck>();
         builder.Services.AddSingleton<IEnvironmentCheck, LegacySettingsFileCheck>();
